requirements.txt
selenium==3.141.0
webdriver_manager==3.2.2
python-telegram-bot==20.8
pyyaml==6.0.1